update code fe

+++ src/main/java/com/example/ecommer/service/CategoryService.java
@@ -19,7 +19,7 @@ public interface CategoryService {
 
     void update(Category category);
 
-    Category findById(Integer id);
+    Category findById(Long id);
 
 //    List<Node> categoryWithIndent();
